ai-server/requirements.txt
ultralytics==8.3.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.9.0.80
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
numpy==1.26.4
roboflow==1.1.34
Pillow==10.3.0
requests==2.32.3
python-dotenv==1.0.1